requirements.txt
playwright==1.42.0
python-telegram-bot==21.0.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4